Interface for ConversationWidget
